fixed rounding bug in intensity calculation found by songshiyu

+++ velodyne_pointcloud/src/lib/rawdata.cc
@@ -221,7 +221,7 @@ namespace velodyne_rawdata
                              * (1 - corrections.focal_distance / 13100);
           float focal_slope = corrections.focal_slope;
           intensity += focal_slope * (abs(focal_offset - 256 * 
-            (1 - tmp.uint/65535)*(1 - tmp.uint/65535)));
+            (1 - static_cast<float>(tmp.uint)/65535)*(1 - static_cast<float>(tmp.uint)/65535)));
           intensity = (intensity < min_intensity) ? min_intensity : intensity;
           intensity = (intensity > max_intensity) ? max_intensity : intensity;
